test(eval): consolidate on-demand regression coverage
